+++ samples/video/hello_video_observers.rb
@@ -5,17 +5,22 @@ include Glimmer
 video_file = File.expand_path('../videos/Ants.mp4', __FILE__)
 
 def display_video_status(video, status)
-  message_box = MessageBox.new(video.swt_widget.getShell)
-  message_box.setText(status)
-  message = "Video Position: #{video.position} seconds\n"
-  message += "Video Duration: #{video.duration} seconds"
-  message_box.setMessage(message)
-  message_box.open
+  message_box {
+    text status
+    message "#{video.position.round(2)}/#{video.duration.round(2)} seconds have elapsed."
+  }.open
 end
 
 @shell = shell {
   minimum_size 800, 500
   @video = video(file: video_file, background: :black) {
+    on_swt_show { |event|
+      # set focus as soon as the SWT widget is shown to grab keyboard events below
+      @video.set_focus
+    }
+    on_key_pressed { |event|
+      @video.toggle if event.keyCode == swt(:space) || event.keyCode == swt(:cr)
+    }   
     on_playing {
       display_video_status(@video, 'Playing')
     }